Server client communication update
The server now sends each connected client a unique hand when the deal
cards button is pressed.

The client receives the hand from the server and adds it to their hand and
draws the cards onto the scene.

Added new hands for each possible player.

Adjusted the scaling of the game to fully scale to fit the entire inner
window of the browser.