Added game logic to help decide which combination of cards are valid to play.
The game logic is checked whenever a player attempts to play a card. Added
a few methods to check if a certain combination is played.

The cards in each player's hands are now sorted in ascending order by the
player before being sent to the player. The sorting algorithm uses quick sort
to sort the cards based on their value and then their suits.

The values for the '2' and 'Ace' playing cards are now adjusted to be higher
than the rest for easier comparisons