Added room creation and lobby list refresh logic
The room creation logic is completed for both the server and client. The
client can initiate the room creation process by clicking the 'Create Room'
button on the lobby screen, and it will create the room on the server-side.
The client will then move on to the next game scene if successful.

The lobby list refresh logic is complete, hitting the refresh button will
update the lobby list of the current game rooms available according to the
server.