Added a users list to the lobby and various additions. See Notes.
A list of users has been added to the lobby screen. Similar to the list of
rooms, the list of usernames shows the people connected to the game.

When a player leaves, their usernames are removed from the scene and they are
removed from the game on the client-side as well. They are also removed
from the turn order (only happens if they disconnect or hit the back button).

No longer allow usernames with special characters: '$#@'"()&%...etc'

Cards now reposition themselves after being played.